Some restructuring of the UpdateManager and an additional force parameter that is run when updates.json is first created
