+++ src/main/kotlin/dev/vxrp/Main.kt
@@ -3,7 +3,8 @@ package dev.vxrp
 import dev.vxrp.configuration.ConfigurationManager
 import dev.vxrp.configuration.data.Config
 import dev.vxrp.configuration.data.Translation
-import dev.vxrp.updates.Updates
+import dev.vxrp.updates.UpdateManager
+import dev.vxrp.updates.handler.UpdateHandler
 import dev.vxrp.util.launch.LaunchOptionManager
 import org.slf4j.LoggerFactory
 
@@ -11,7 +12,7 @@ private val logger = LoggerFactory.getLogger("Main")
 
 fun main() {
     logger.info("Starting up...")
-    val updates = Updates()
+    UpdateManager().checkUpdated()
 
     val configurationManager = ConfigurationManager()
 
@@ -21,7 +22,7 @@ fun main() {
     configurationManager.initializeDatabase(config)
     configurationManager.setLoggingLevel(config)
 
-    updates.checkForUpdatesByTag(config, "https://api.github.com/repos/Vxrpenter/SCPToolsBot/git/refs/tags")
+    UpdateHandler().checkForUpdatesByTag(config, "https://api.github.com/repos/Vxrpenter/SCPToolsBot/git/refs/tags")
     ScpToolsBot(config, translation)
 }
